[uni] completely restructure the project
refactor into lexer, parser and, interpreter modules
