+++ src/lexer/mod.rs
@@ -1,0 +1,8 @@
+mod lexer;
+pub mod token;
+
+pub use lexer::Lexer;
+
+#[cfg(test)]
+mod tests;
+
