+++ src/main.rs
@@ -1,6 +1,6 @@
 #![allow(dead_code, unused_variables)]
 
-use std::{io};
+use std::io;
 use std::io::{ErrorKind, Read};
 use std::time::Instant;
 use clap::Parser;
@@ -8,15 +8,12 @@ use miette::{miette, Result};
 
 use crate::aplang::ApLang;
 use crate::arguments::{CommandLine, DebugMode};
-use crate::errors::Reports;
+use interpreter::errors::Reports;
 
-mod ast;
-mod errors;
 mod interpreter;
 mod lexer;
 mod parser;
-mod token;
-mod aplang_std;
+mod standard_library;
 mod arguments;
 mod aplang;
 
@@ -28,41 +25,6 @@ fn main() -> Result<()> {
 
 fn run(args: CommandLine) -> Result<()> {
     let mut debug_buffer = String::new();
-
-    // note: consider adding debug logs here
-    // load the source code
-    // let source_code: Arc<str> = if let Some(file_path) = &args.file {
-    //     if args.enforce_file_extension {
-    //         let ext = file_path.extension()
-    //             .ok_or(miette!("could not read file extension from file"))
-    //             .map(|os_str| os_str.to_string_lossy().into_owned())?;
-    //
-    //         if ext.eq_ignore_ascii_case(".ap") {
-    //             return Err(miette!("file extension is not '.ap' instead found .{}", ext))
-    //         }
-    //     }
-    //
-    //     file_name = Some(file_path);
-    //
-    //     // file_name = file_path.file_name()
-    //     //     .map(|os_str| os_str.to_string_lossy().into_owned())
-    //     //     .ok_or(miette!("failed to read filename from file"))?;
-    //
-    //     fs::read_to_string(file_path).map_err(|error| miette!(
-    //        "failed to open file {:?}\n{}", file_path.as_path(), error
-    //     ))?.into()
-    //
-    // } else if let Some(eval) = args.eval {
-    //     eval
-    // } else if args.eval_stdin {
-    //     let mut buffer = String::new();
-    //     io::stdin().read_to_string(&mut buffer).map_err(|error| miette!(
-    //         "failed to read from stdin\n{}", error
-    //     ))?;
-    //     buffer.into()
-    // } else {
-    //     unreachable!()
-    // };
 
     let aplang = if let Some(file_path) = args.file {
         ApLang::new_from_file(file_path.clone()).map_err(|err| {
@@ -88,8 +50,6 @@ fn run(args: CommandLine) -> Result<()> {
         unreachable!()
     };
 
-    // let aplang = ApLang::new_from_stdin(source_code, file_name);
-    
     // execute the lexer
     let lexed = aplang?.lex().map_err(Reports::from)?; // todo implement errors here
     
@@ -106,7 +66,6 @@ fn run(args: CommandLine) -> Result<()> {
         .map_err(Reports::from)?; 
     
     // todo implement errors here
-    
     if matches!(args.debug, DebugMode::All | DebugMode::Parser) {
         parsed.debug_output(&mut debug_buffer).map_err(|err| {
             miette!("could not write debug info for parser!\n{}", err)
