+++ src/standard_library/file_system.rs
@@ -1,10 +1,11 @@
 use std::cell::RefCell;
 use std::fs;
-use std::fs::{OpenOptions};
+use std::fs::OpenOptions;
 use std::io::Write;
 use std::path::Path;
 use std::rc::Rc;
-use crate::interpreter::{FunctionMap, Value};
+use crate::interpreter::FunctionMap;
+use crate::interpreter::Value;
 use crate::std_function;
 
 pub(super) fn file_system() -> FunctionMap {
