+++ src/standard_library/strings.rs
@@ -1,6 +1,7 @@
 use std::cell::RefCell;
 use std::rc::Rc;
-use crate::interpreter::{FunctionMap, Value};
+use crate::interpreter::FunctionMap;
+use crate::interpreter::Value;
 use crate::std_function;
 
 pub(super) fn std_strings() -> FunctionMap {
